Added breakpoint for medium-sized screens
I copied my css into the breakpoints themselves, hopefully creating a more code-efficient and simpler design. I then tweaked the code so it would work for both desktop and tablet. Next, I will recode the mobile layout. This one took a while to do, not because it was difficult, but because of the programmer's own stupidity.