Finished navbar for the other breakpoints, changed sizing units
This one was also late due to being behind in other subjects, but I redid the code for tablet and desktop layout and changed all the size units for text to be in em rather than % or vw, which could be messy with my responsive layout. I wanted a semi-consistent design, so I decided to keep the font scaling the same through the breakpoints, (eg 1em stayed 1em, but the size of the text would still scale on the final site.) at this point, the text looks fine, but I might make the text bigger on desktop later depending on the final site. Again, I don't really like the colour scheme, but I want to finish my page before I tweak it (I'm learning what minimum viable product is firsthand lol)