Added body for mobile breakpoint
I was going to split this into two seperate commits, but I went on and finished the whole section anyway because I didin't want to make too many commits. I used the CSS Grid layout to put the elements in the body. I used 3 fonts from google fonts, Monda, Share Tech, and Source Sans 3. I had used Source Sans Pro on the design, but I couldnt get it easily and this one was easy, not to mention more browser-friedly as it is a common font. I then used div elements to create circular icons. There was a problem with adding a link to them, because putting them in an <a> somehow removed the parent tag from the div, causing the styles to be removed. I have left this problem for now, because there is probably now way to fix it with the way I have been doing it. I also realised an earlier oversight with my CSS grid (who would have guessed) and forgot about the subtitles for the icon links. so added new rows into by body tag. I used a smaller fr value to align my elements easier. I also changed a Lot of names back and forth to make my code more optimised, so I used both id and class to cut down on my code size. I've learned my lesson from last week :P