Medium screen layout
I changed the css for the tablet layout to fit my medium screen layout. I also tweaked the body tag so that the footer floats to the bottom, also known as a sticky grid. I looked at this tutorial : https://www.youtube.com/watch?v=csQkSrw1Ghk to do this, but it was so basic that I could have just wrtitten the four lines of code without wasting four minutes, 0/10 would not reccomend. The body tag is now a grid, and the header, main, and footer are the container items, so I could tweak the grid row sizings. I also trialled the column sizings (on trello).