Finished layout for homepage for larger screens
Copied code over to the larger screen layout. Feedback on layout to follow after this. Forgot this on my last commit, I added basic code for my footer's modal about last week, but I ran into an issue and need to fix it. it's on the back burner for the moment, and i'll probably replace it later.