End-User Feedback Round 1
This change is detailed in my Trello Board. I made a list of things that I needed to change, then implemented them iteratively. I am also working on my code optimisation, as is shown in the top of my CSS code. More to come in the next commit, I just want to save it for now. I also did a lighthouse test, and the results are on trello as well.