Finished footer and styled the button
This took place during the T2 holidays, so I didin't get much done on the project. I didin't want to put the project off, so I finished the footer, which was
relatively basic at the moment. I want to add more styles in once I actually had my assets in my site, so I can credit them. The tutorial I used had a button, which works onclick. I wanted to style it like a link though, so I removed the background and added a hover function. The feedback form links to a google form that I hope to keep updated after my project is implemented.