Copied header and footer over to heuristics.html and added title and paragraph
I copied my code over to my next page, and used the pre-existing title and paragraph to place the title and paragraph. In this page, the subtitle is not there, so I did a little bit of messing around with my code so it appears in the second row instead of the first. I tried to redo most of my code but ended up using a unique id so it selected only the one in the heuristics page.