Accordion - Iteration 1
This was the first iteration of my collapsing accordion menus. I got them to display underneath each other and show their content when clicked, but on the large and medium layout, my end-users said that having them all in one row looked too overwhelming. I wanted to display them in a 2x5 rather than a 1x10, so I tried to get them to display underneath each other. This design did not work with my css grid, because the menu and items are seperate classes.